create deletePost() in models/Post.php

+++ SharePostsMVC/app/controllers/Posts.php
@@ -120,6 +120,7 @@ class Posts extends Controller {
             if($post->user_id != $_SESSION["user_id"]) {
                 redirect("posts");
             }
+            
             $data = [
                 "id" => $id,
                 "title" => $post->title,
@@ -132,6 +133,14 @@ class Posts extends Controller {
     
     public function delete($id) {
         if($_SERVER["REQUEST_METHOD"] == "POST") {
+            // Get existing post from model
+            $post = $this->postModel->getPostById($id);
+            
+            // Check for owner
+            if($post->user_id != $_SESSION["user_id"]) {
+                redirect("posts");
+            }
+            
             if($this->postModel->deletePost($id)) {
                 flash("post_message", "Post Removed");
                 redirect("posts");
